up: improve jumbo line scrolling by moving up from eol
The special exception which defeats re-centering occurs when the prior
point (not line) is on a over-tall jumbo. Since there could be text
before the jumbo image, we move first to the end of visual line, since
that's full height (except in visual-line-mode, where thankfully the
exception seems to apply to the full line).